+++ ui/src/component/App.tsx
@@ -1,20 +1,18 @@
 import "./App.css";
 import React from "react";
 import { Routes, Route } from "react-router-dom";
-import ProfileView from "../views/ProfileView";
-import LoginView from "../views/LoginView";
-import GameView from "../views/GameView";
-import { makeStyles } from "tss-react/mui";
+import ProfileView from "../views/Profile/ProfileView";
+import LoginView from "../views/Login/LoginView";
+import GameView from "../views/Game/GameView";
 import { RoutePath } from "../interfaces/router.interface";
 import { TranscendanceContext } from "../context/transcendance-context";
 import { TranscendanceStateActionType } from "../context/transcendance-reducer";
 import Toast from "../context/toast";
+import { Box } from "@mui/material";
 
 // TO DO : set up theme for these colors #1d3c45, #d2601a, #fff1e1
 
 export default function App() {
-  const { classes } = useStyles();
-
   function closeToast() {
     dispatchTranscendanceState({
       type: TranscendanceStateActionType.TOGGLE_TOAST,
@@ -35,24 +33,20 @@ export default function App() {
           onClose={closeToast}
         />
       )}
-      <div className={classes.root}>
+      <Box
+        sx={{
+          display: "flex",
+          flexDirection: "column",
+          minHeight: "100vh",
+          backgroundColor: "#fff1e1",
+        }}
+      >
         <Routes>
           <Route path={RoutePath.LOGIN} element={<LoginView />} />
           <Route path={RoutePath.PROFILE} element={<ProfileView />} />
           <Route path={RoutePath.GAME} element={<GameView />} />
         </Routes>
-      </div>
+      </Box>
     </>
   );
 }
-
-// https://github.com/garronej/tss-react
-
-const useStyles = makeStyles()(() => ({
-  root: {
-    display: "flex",
-    flexDirection: "column",
-    minHeight: "100vh",
-    backgroundColor: "#fff1e1",
-  },
-}));
